Estimate a low-rank network model from Aggregated Relational Data (ARD)

Estimate a low-rank network model from Aggregated Relational Data (ARD)

Usage

estimate_spectrum(ard, traits)

Arguments

ard

A matrix containing Aggregated Relational Data. Each row of ard corresponds to one node in the network, and each column represents an aggregated trait. The order of the nodes (rows) and aggregated traits (columns) should match traits. Missing data is not allowed.

traits

A matrix containing node-level trait information. Each row of traits corresponds to one node in the network, and each column represents a particular trait. The order of the nodes (rows) and aggregated traits (columns) should match ard. Missing data is not allowed.

Value

A SpectralArdEstimate object.

Examples

sim <- simulate_ard_data(n = 100, k = 2, corr = 0.7,
  num_good_traits = 2, num_bad_traits = 2)

estimate <- estimate_spectrum(sim$Y, sim$traits)
estimate
#> $Utilde
#>                 [,1]          [,2]          [,3]          [,4]
#>   [1,] -1.914514e-01  0.0150572486 -0.1478340767 -0.0192574345
#>   [2,] -2.052871e-01  0.0388356985  0.0680789706 -0.0008784947
#>   [3,] -1.928169e-01 -0.0715049254 -0.1826832462  0.0300581874
#>   [4,] -1.675666e-01 -0.0701219465 -0.0003175608  0.0924187505
#>   [5,] -2.021652e-01  0.0417399790  0.0149132569 -0.1358355710
#>   [6,] -1.700788e-01 -0.0330202724  0.1480760189  0.0209298320
#>   [7,] -1.700226e-01 -0.0137266928  0.1869325190  0.1314096703
#>   [8,] -1.618294e-01 -0.0234223819  0.0525795962 -0.0482345328
#>   [9,] -1.596480e-01 -0.0546944669 -0.0027622347 -0.1040573523
#>  [10,] -1.383713e-01 -0.1455144041 -0.2587780839  0.0085367392
#>  [11,] -1.506191e-01  0.0409232846  0.0196032266 -0.1153973225
#>  [12,] -1.340189e-01 -0.0170174084 -0.0013189766  0.0735232180
#>  [13,] -1.574471e-01 -0.1086946568 -0.1111574911  0.0421007382
#>  [14,] -1.196923e-01  0.0133892334  0.0126007991  0.0013791661
#>  [15,] -1.430397e-01  0.0214380257  0.1111088801 -0.2057691173
#>  [16,] -1.448015e-01  0.0323888265  0.0378502821 -0.1650903887
#>  [17,] -1.229775e-01 -0.0215480358 -0.0977455231 -0.0096007664
#>  [18,] -1.259927e-01 -0.0564840474 -0.0003128080 -0.0188752994
#>  [19,] -1.520450e-01 -0.0458903651  0.0527649040  0.2288938595
#>  [20,] -1.764265e-01  0.0271697022  0.1613327168  0.1228061336
#>  [21,] -1.224869e-01  0.0313999327 -0.0300677209 -0.0731464243
#>  [22,] -1.450880e-01 -0.0641740443 -0.0167015700  0.0156413426
#>  [23,] -1.036896e-01 -0.0030380406 -0.0346300595 -0.0551261899
#>  [24,] -1.105672e-01 -0.0335982116 -0.0400392202 -0.0373963245
#>  [25,] -9.801128e-02  0.0074905493  0.0853434741 -0.0264370156
#>  [26,] -9.079392e-02 -0.0119683538 -0.0489613932  0.0154244936
#>  [27,] -1.143653e-01  0.0396706142  0.1636709695 -0.1407321775
#>  [28,] -1.101431e-01 -0.0103726363 -0.0891188440 -0.0565039247
#>  [29,] -9.719251e-02 -0.0538848435 -0.1177643231 -0.0331768875
#>  [30,] -1.070577e-01 -0.0391834433 -0.1367222793 -0.0438373230
#>  [31,] -1.364212e-01 -0.0645402237  0.0005716734  0.0019345102
#>  [32,] -1.532923e-01  0.0007987165  0.0384357283 -0.0518023743
#>  [33,] -1.124190e-01  0.0181902774  0.0693245633  0.0368251222
#>  [34,] -1.610482e-01  0.0355108269  0.1812342323 -0.0570158108
#>  [35,] -1.354929e-01 -0.0185867076  0.1079277309  0.1647382109
#>  [36,] -1.227409e-01 -0.0022315126  0.1038297213  0.0163447929
#>  [37,] -1.255611e-01  0.0167048044 -0.0279427780 -0.0436147632
#>  [38,] -1.209016e-01 -0.0590848579 -0.1214516368  0.0394416726
#>  [39,] -1.375568e-01 -0.0362066961  0.1599617129  0.1977782487
#>  [40,] -8.176743e-02 -0.0444217096 -0.0708334985  0.0450771211
#>  [41,] -8.465172e-02 -0.0151907307 -0.1681192733 -0.0001339920
#>  [42,] -1.041932e-01  0.0033203051  0.1600183012  0.0226020301
#>  [43,] -9.498257e-02  0.0649678200 -0.0361080159 -0.3080419225
#>  [44,] -8.294105e-02  0.0037531127 -0.0449938279 -0.1276217665
#>  [45,] -7.121097e-02  0.0201609825  0.0727647288 -0.0572805914
#>  [46,] -8.822487e-02 -0.0475160966 -0.0610392937  0.1053598829
#>  [47,] -5.765835e-02 -0.0273583763 -0.1090622849 -0.1601645874
#>  [48,] -7.559084e-02 -0.0549794345 -0.0720951196  0.0132600910
#>  [49,] -9.493927e-02 -0.0490006182 -0.1741678578  0.0152961508
#>  [50,] -8.894294e-02 -0.0384396926  0.0971092537  0.0921369510
#>  [51,] -1.083936e-01  0.0266955932  0.0064598206 -0.0710691054
#>  [52,] -9.921007e-02 -0.0421205217 -0.0771745045  0.1105561948
#>  [53,] -1.211306e-01 -0.0774126191 -0.0187284066  0.1133152572
#>  [54,] -9.674594e-02 -0.0856529556 -0.1569213150  0.1736160913
#>  [55,] -7.195634e-02  0.0268070225  0.0202522352 -0.0858359514
#>  [56,] -6.912741e-02 -0.0298667505 -0.1985800364 -0.0685521837
#>  [57,] -1.177805e-01 -0.0069235158  0.0744341245 -0.0631771592
#>  [58,] -3.366020e-02 -0.0245456718 -0.1436068211 -0.1282107483
#>  [59,] -7.148572e-02 -0.0409550479  0.0630608689 -0.1268089084
#>  [60,] -6.322495e-02 -0.0400186751  0.1570605291  0.0577771422
#>  [61,] -7.100267e-02  0.0186449731 -0.0324108070 -0.1626321987
#>  [62,] -4.696788e-02  0.2445600658  0.0048075699  0.1014176540
#>  [63,] -4.533773e-02  0.1841543502 -0.2630315425  0.0928056212
#>  [64,] -4.334292e-02  0.2574418432  0.0659775892 -0.1592501142
#>  [65,] -4.233883e-02  0.1606580989 -0.1100653785  0.1139599214
#>  [66,] -2.388642e-02  0.0612715766 -0.1619334928  0.1845726792
#>  [67,] -6.237254e-02  0.1542594752  0.0674175894  0.3248848188
#>  [68,] -3.272630e-02  0.1229587179 -0.0178391999  0.1344316029
#>  [69,] -1.179578e-02  0.1300386555 -0.0619588100  0.0625511911
#>  [70,]  6.419167e-03  0.2161222026 -0.0148767998 -0.1073098359
#>  [71,] -8.049757e-03  0.1694642299 -0.0393274387 -0.0813832668
#>  [72,] -2.295070e-02  0.1360881161 -0.0038927788  0.0965825475
#>  [73,] -3.325170e-02  0.1604543947  0.1003225294  0.1307553861
#>  [74,] -6.506217e-03  0.1309603881 -0.0227812705 -0.0075027448
#>  [75,]  1.075540e-02  0.1304589751 -0.1335892474  0.0566708699
#>  [76,] -2.976493e-02  0.2294483801 -0.0527077338 -0.0759674170
#>  [77,] -4.048128e-02  0.1224255202  0.1116892776  0.0775263868
#>  [78,] -2.123421e-02  0.1720966690 -0.1209168879 -0.0654328269
#>  [79,] -2.165957e-02  0.0335778418 -0.0351446770  0.0691844158
#>  [80,] -1.716056e-02  0.2488938928 -0.0399668364 -0.0398444508
#>  [81,] -5.143076e-02  0.1495946609  0.1031802124  0.0140881805
#>  [82,] -1.902820e-02  0.1242592679 -0.0419150943 -0.0280336235
#>  [83,] -3.143212e-02  0.0851626459 -0.0495140940  0.1200607533
#>  [84,] -2.467524e-02  0.0805372312 -0.0022213644  0.0029721203
#>  [85,] -1.865576e-02  0.1782038913  0.0115011970 -0.0799739065
#>  [86,] -2.245983e-02  0.1203598868 -0.1298540959  0.0415852136
#>  [87,]  1.037487e-05  0.0761905894 -0.0825479736  0.0150887682
#>  [88,] -1.887034e-02  0.1461812944 -0.0074001015 -0.0105469722
#>  [89,] -1.946036e-02  0.2068548168  0.0642389344 -0.0418062877
#>  [90,] -7.086943e-03  0.0790354908 -0.0468638257 -0.0439109285
#>  [91,] -1.843584e-02  0.1567931156 -0.1578818159  0.0014898802
#>  [92,] -1.939130e-02  0.0612510395 -0.1385183599  0.1115777862
#>  [93,] -9.979002e-03  0.1249688149 -0.0489904393  0.1432345920
#>  [94,] -3.472702e-02  0.1111233415  0.0730856273  0.0420379432
#>  [95,] -1.561129e-02  0.1049249521 -0.0747752624  0.0725996049
#>  [96,] -2.000024e-02  0.0420464973  0.0834285738  0.0049270354
#>  [97,] -1.213630e-02  0.1419818836 -0.0571308874  0.0213293109
#>  [98,] -3.193304e-03  0.0622066586 -0.0064463710  0.0987890295
#>  [99,] -1.215398e-02  0.0547505481 -0.0358914314 -0.0132444704
#> [100,] -3.103171e-02  0.2159184576  0.1108635586 -0.0127206939
#> 
#> $Sigmatilde
#>           [,1]       [,2]       [,3]       [,4]
#> [1,] 28.077783  -2.129055   2.010274   4.172625
#> [2,] -2.129055  26.768108  11.447016 -11.433404
#> [3,]  2.010274  11.447016  10.718129 -24.904723
#> [4,]  4.172625 -11.433404 -24.904723  17.580711
#> 
#> attr(,"class")
#> [1] "SpectralArdEstimate"